Use -T1 for xz but respect NJOBS in more bzip3+kanzi cases
Force xz single-threaded with -T1 to get exactly the same size results
with recent xz versions (> 5.4.x) as with xz versions <= 5.4.x.

Use NJOBS as number of threads for more of the initial non-parallel
bzip3 and kanzi cases since this doesn't affect size at all, won't use
too much memory and may make the script a bit faster for large files.

Write "Usage" message to stderr instead of stdout, more in line with
tradition and more visible when output redirected to a file.

Reorder transform list for the 2- and 3-level loops to get more
interesting results early on in some cases.